bomberman/requirements.txt
websockets>=12.0
aiortc>=1.6.0
pygame>=2.5.0
psycopg2-binary>=2.9.0
python-socketio[client]>=5.11.0
pyperclip>=1.8.2

